coinview/requirements.txt
python-binance
TA-Lib
numpy
backtrader
decouple
flask